+++ Example/DLRadioButtonExample/DLDemoViewController.h
@@ -3,6 +3,4 @@
 
 @interface DLDemoViewController : UIViewController
 
-@property (weak, nonatomic) IBOutlet DLRadioButton *waterButton;
-
 @end