+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [1.4.8](https://github.com/DavydLiu/DLRadioButton/releases/tag/1.4.8) (04/10/2016)
+* Added example for using with Swift.
+* Made APIs more Swift friendly.
+
+
 ## [1.4.7](https://github.com/DavydLiu/DLRadioButton/releases/tag/1.4.7) (04/02/2016)
 * Added support for RTL interface layout.
 * Updated API comments.